hf_space/requirements.txt
tensorflow==2.15.1
gradio==4.44.1
numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84
pandas==2.2.2
